riscv: mm: dma-noncoherent: Switch using function pointers for cache management
Currently, selecting which CMOs to use on a given platform is done using
the ALTERNATIVE_X() macro. This was manageable when there were just two
CMO implementations, but now that there are more and more platforms coming
needing custom CMOs, the use of the ALTERNATIVE_X() macro is unmanageable.

To avoid such issues this patch switches to the use of function pointers
instead of the ALTERNATIVE_X() macro for cache management.

    void (*dma_cache_wback)(void *vaddr, unsigned long size);
    void (*dma_cache_inv)(void *vaddr, unsigned long size);
    void (*dma_cache_wback_inv)(void *vaddr, unsigned long size);

The above function pointers are provided to be overridden for platforms
needing CMO.

Convert ZICBOM and T-HEAD CMO to use function pointers.

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>