arm64: dts: renesas: rzt2h-n2h-evk: Enable Ethernet switch
Enable the Ethernet switch on the RZ/T2H and RZ/N2H EVK by activating the
ethsw node, defining its MDIO bus, and adding PHY0 and PHY1 with their
associated LED configurations.

Enable GMAC0 for use as the CPU port. Reduce the number of TX queues to
accommodate the MTU restrictions when GMAC0 is used as a DSA CPU port.

Mark the MIIC converters connected to the switch ports as active and
configure phylink polarity. Add switch port nodes for lan0 and lan1 with
the appropriate PHY handles and enable switch port 3.

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>